code/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.8.0
pydantic-settings>=2.4.0
sqlalchemy>=2.0.0
PyJWT>=2.9.0
cryptography>=42.0.0
python-multipart>=0.0.9
